refactor: remove anthropic messages bridge path
